requirements.txt
streamlit
google-generativeai
PyMuPDF
python-dotenv
Pillow
plotly
pandas
